fix(ghosts-alpha): wrong gclient_s struct that was preventing the common ufo implementation from working

+++ src/Games/AlphaGhosts/Structs.h
@@ -33,7 +33,7 @@ struct gclient_s
 {
     playerState_s ps;
     char padding1[0x360];
-    uint8_t mFlags;
+    int mFlags;
     char padding2[0x41C];
 };
